+++ flutter/example/lib/screen/demo/search_example/search_screen.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_remote_ui/flutter_remote_ui.dart' as rui;
 
 class SearchScreen extends StatefulWidget {
   static const routeName = "/demo/search";
@@ -12,8 +13,9 @@ class _SearchScreen extends State<SearchScreen> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text("data"),
+        title: Text("search result"),
       ),
+      body: buildBody(),
     );
   }
 
@@ -31,16 +33,39 @@ class _SearchScreen extends State<SearchScreen> {
   Widget buildSearchResults() {
     return ListView.builder(
       itemBuilder: (c, i) {
-        return SearchResultRow();
+        return SearchResultRow(
+          icon:
+              Icon(rui.RemoteIconData.fromUri("material://Icons.person").icon),
+          title: Text("data"),
+          subtitle: Text("sub"),
+          onTap: () {},
+        );
       },
       itemCount: 12,
+      shrinkWrap: true,
+      physics: NeverScrollableScrollPhysics(),
     );
   }
 }
 
+@rui.LayoutConvert("search-result-row")
 class SearchResultRow extends StatelessWidget {
+  final Widget title;
+  final Widget subtitle;
+  final Widget icon;
+  final Function onTap;
+
+  const SearchResultRow(
+      {Key key, this.title, this.subtitle, this.icon, this.onTap})
+      : super(key: key);
+
   @override
   Widget build(BuildContext context) {
-    return Container();
+    return ListTile(
+      onTap: onTap,
+      leading: icon,
+      title: title,
+      subtitle: subtitle,
+    );
   }
 }