+++ flutter/lib/flutter_remote_ui.dart
@@ -5,6 +5,7 @@ export 'src/images/images.dart';
 export 'src/layouts/layouts.dart';
 export 'src/action/action.dart';
 export 'src/prebuilt/prebuilt.dart';
+export 'src/annotations/annotations.dart';
 
 // region default packages
 export 'package:flutter_remote_icon/flutter_remote_icon.dart';