+++ flutter/lib/src/annotations/layout.dart
@@ -1,0 +1,5 @@
+class LayoutConvert {
+  final String name;
+
+  const LayoutConvert(this.name);
+}