+++ flutter/packages/flutter_remote_icon/CHANGELOG.md
@@ -1,3 +1,18 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.1] - Initial release
 
-* TODO: Describe initial release.
+## [0.1.0] - (in development)
+
+RemoteIconData as factory or as `IconData`
+```dart
+// from 
+Icon(RemoteIconData.fromUri("~").icon);
+// to 
+Icon(RemoteIconData.fromUri("~"));
+
+
+// from
+RemoteIcon(icon: RemoteIconData.fromUri("~").icon);
+
+// to 
+RemoteIcon.fromUri("~");
+```